+++ app/models/__init__.py
@@ -1,3 +1,4 @@
+from app.models.chat import AddedChat
 from app.models.user import User
 
-__all__ = ["User"]
+__all__ = ["User", "AddedChat"]